+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog (https://keepachangelog.com), and this project adheres to Semantic Versioning (https://semver.org).
 
+## [Unreleased] - 2026-02-18
+### Fixed
+- Fix a bug when testing strings for equality.
+- Add a small basic game `nicomas.bas`.
+- Add operator unit tests.
+
 ## [Unreleased] - 2026-02-17
 ### Added
 - Add `RIGHT$()` support in Apple II Basic. Add relevant unit tests.